+++ src/Connection.php
@@ -260,6 +260,11 @@ class Connection extends Component
      */
     public $hostname = 'localhost';
     /**
+     * @var string the connection scheme used for connecting to the redis server. Defaults to 'tcp'.
+     * @since 2.0.18
+     */
+    public $scheme = 'tcp';
+    /**
      * @var string if the query gets redirected, use this as the temporary new hostname
      * @since 2.0.11
      */
@@ -593,7 +598,7 @@ class Connection extends Component
             return 'unix://' . $this->unixSocket;
         }
 
-        return 'tcp://' . ($this->redirectConnectionString ?: "$this->hostname:$this->port");
+        return $this->scheme . '://' . ($this->redirectConnectionString ?: "$this->hostname:$this->port");
     }
 
     /**